Refactor : 컴포넌트 네이밍 문법 변경
상세 : airbnb style guide에 따라 파일 이름을
파스칼 형식으로 변환
