+++ client/src/presentation/components/index.js
@@ -1,3 +1,3 @@
-import Logo from './logo';
+import Logo from './Logo';
 
 export default Logo;
